Restructure the community section
The community section included resources for both the community, and
development standards. These are both unversioned items, which do not
belong in the version-specific documentation.

To better organise these, I've renamed the 'community' to 'general', and
then organised the existing documentation within that section.

I've generated separate sidebars for the general folder, and linked to
the initial content for each via the navbar.

All existing links are updated, and a couple of new entrypoints have
been introduced.

I've also moved the Develop Meetings into /general/community/meetings.

At this point I have _not_ renamed the meetings to 'Developer meetings'
or similar. That should probably be a separate PR once a decision is
made on that content.