Link to courses on the Moodle Academy
This is part of a change to introduce standard links to courses on
Moodle Academy.

We create a single canonical list of courses in `/academyCourses.js`,
a custom Admonition via a JSX component, and a custom page listing all
academy courses relevant to Development.

That component can take properties to describe:

- the topic being taught
- a short name for the course
- possibly further information to include

The list of courses in academyCourses.js has a list of the courses, with
their shortname as an object key, and then relevant metadata, including:

- The course full name
- The course link or id
- A summary to be used in a tooltip
- A full description to be used in the course list page
- information as to whether the course has been archived
- tags
- icon

The academyCourses page also uses this course information to list all
courses available, with further information and tags.

Fixes #12