[system] Keep track of migrated docs and warn on use
The commit adds a new migratedPages.js module which includes a list of
legacy page to new page mappings.

When a legacy link in the Wikimedia internal link format is encountered,
it is checked against this list of migrated pages, and an updated page
URL is generated.

This value is notified to the Docusaurus CLI console, and the updated
link is used in the generated documentation.