[docs] Add Testing docs
This commit adds the following testing docs:

- Testing
- Testing integrated issues
- QA testing

I've also updated the content of these pages:

- I've removed references to Wednesday (as testing day)
in the Testing integrated issues, because we're not
following this approach anymore.
- I've added reference and link to Behat in Testing page.

The testing sidebar now is generated manually, because
the items should be displayed in the same order as they
appear in the workflow.

Apart from that, some of the testing docs linked from Testing
page, such as Testing integrated issues or QA testing  won't be
displayed in the sidebar (but the sidebar will be displayed
when this doc is accessed).