[system] Add a hacky way to hide docs from sidebar
In some pages, we might want to exclude docs from sidebar
generation (but display sidebar when accessing to those
docs).
We've been testing the [ref" item](https://docusaurus.io/docs/next/sidebar/multiple-sidebars#sidebar-item-ref)
but it's for a different purpose.

So, for now, we've added the 'docusausus-hidden' class
to the sidebar in order to display a sidebar for these pages,
without displaying any link in the sidebar.

Example of use:

```
{
    type: 'doc',
    id: 'development/process/testing-integrated-issues',
    className: 'docusaurus-hidden',
},
```