[system] Split Process and Policies in navbar
Instead of having "Process & Policies" in the navbar, I suggest
having 2 separate items for them. That way, we'll be able to
have more levels in the pages and users will find this
information easily.

For now, the patch also moves the "Development" page inside
Process (because it's part of it).

There is no Policies default page so, for now, I've linked
the pre-existing Coding style.
In the future, we can add a separate page or try to add more
information to this category...

But after doing this PoC, I really like how the "Process"
category looks like.