[repo] Create a basic fetch script
This is a really basic system to fetch a page from the legacy docs into
a new location on disk, and perform basic changes.

It does *NOT* catch everything and it can do better, but it's what I've
been doing manually.

At the moment it will:

- not convert any table
- complain about migrated pages but not fix them
- do nothing with tracker links

I've tried to merge some of the functions and utilities that I
previously wrote for the legacy wiki updater into a shared utils file
too.

Co-authored-by: Sara Arjona <sara@moodle.com>