[repo] Rewrite the migration script to use markdownlint
The original migration was flaky and based on regex. It was not syntax
aware, and cauesd things like code sections to break.

This new approach makes use of Custom markdownlint rules with
markdownlint-cli2 to make incremental syntax-aware changes to the
markup.

This means that things like `=` and `#` in code blocks do not break.

This still does not introduce a solution for converting uses of
wikimedia templates, or wikimedia tables in the source content, and this
may not be easily possible whichever approach we take, but it does make
most conversions work much better.