[repo] Switch to markdownlint-cli2
This related change migrates from markdownlint-cli to markdownlint-cli2.

Despite the terrible name, markdownlint-cli2 is a more appropriate tool.
The author is the author of the markdownlint library, and it allows us
to specify custom rules _in_ the configuration rather than requiring that
they be specified as CLI arguments making it more natural.

This commit adds/replaces some yarn commands:
- `mdlint [glob] [globn]` - Lint all files mentioned in the globs
- `mdlint-all` - Lint all markdown files
- `mdfix [glob] [globn]` - Fix all mdlint issues in the specified files
- `mdfix-all` - Fix all markdown files