[docs] Migrate releases docs
This change also moves all version data for all released versions to a
JSON structure containing all relevant information for all Moodle
releases since 1.4.5. This is validated using a json-schema.

New React elements are introduced for:

- `ReleaseTable` - to list all releases for a major release
- `SupportedReleases` - to list and summarise only the supported
  releases

All utility functions have been moved to a new `utils/SupportedReleases`
file defining the types and providign useful helpers.

The sidebar configuration has been updated to use the new data source,
but we need to improve general infrastructure to allow the docusaurus
configuration to load typescript files before it can be updated to use
the new utility. That will be investigated in a separate issue.