[docs] general/process: Add apidoc build changes
One week before a major release we should update the list of branches in
both of the apidoc repositories to only include the versions that will
be supported after the release.

Jobs only run on a Sunday evening, so as long as the schedule is
followed, the new documentation will only be generated after the branch
is created for the new version, and the outgoing version will be
documented until a few hours before the release.