[repo] Modify the doc footer to float
This means that, on wider screens, the 'Edit this page' link is always
visible.

I've tried to keep it relatively small and out-of-the way so as not to
detract from content or take up too much space.

I dare say we can improve this in the future, and when we get it working
nicely it would be good to release it as a separate docusaurus plugin.