[docs] docs: Tidy up the Guides landing
This commit adds a new guides landing page for the non-API guides, and
updates the main landing page for the version-specific documentation.

I've also removed the link to the 'Getting started' from the sidebar
because it takes you away from the API information and changes the
sidebar navigation completely, which makes it harder to go back to the
previous area.

As part of this I've also restructured the sidebar to move the Developer
update notes to the top section above the break, so that all
documentation for the version is together, in one place.