[repo] Adjust link selector
The link selector only looked for the first link on the page.

Given the line:

Example [Foo bar](/baz) and this extra [link](https://example.com) here.

It was only fetching the following into the match:

description: Foo bar](/baz) and this extra [link
link: https://example.com

The description expression was too greedy (.*). This change updates the
regex to look for any character which is not a ] in the description.