docs: Add guidance on setting custom task data and user
When we wrote the task API we goofed by allowing just anyone set custom
data. By setting custom data outside of the task consuming the data, we
make the API harder to enforce:
* the parameters may change but we'd never find out without combing
  through logs
* formatting of the custom data is left to the caller

We should really encourage use of a factory to create the task itself.

I think that scheduling should still happen outside of the factory, but
the user configuration also belongs in the factory.