Move the "raise priority to major" task to -5w
This task is to be done the week after freeze. Since recently
we moved the Freeze week to be -6w (from -5w). So this task
needs to be done -5w (from -4w).

The task is documented in the "Standard messages" internal doc
("Raising priority of some minor issues to major" section) and
it has been already amended there.