Document the use of @var also for class constants
The use of @const is incorrect and we have to explicitly say it.
Following both the phpDocumentor guidelines and the phpdoc-tags
draft proposal, @var should be used for both class properties and
constants (and other things that are out from the scope of this change).

This is part of https://tracker.moodle.org/browse/MDL-77599