[docs] Remove registration changes
We were able to deprecate the modal_factory entirely in 4.3, so for new
code aimed at 4.3 onwards, there is no need to register. The new
registration process is only available for 4.3 onwards, so it no longer
makes sense to document this.

Code aimed at versions including 4.2 and older will still need to
register using the previous mechanism.