Peer review checklist: make accessibility checks consistent with others
Accessibility should not be optional in the peer review checklist.
It is something Moodle takes seriously.

On the other hand, the reasons this checklist item was made optional,
was that it was written in a totally aspirational way, which was more
of an accessibility test, and not a peer review.

Therefore, I propose we re-word to acutally be appropriate to be part
of the peer review checklist. There is already a separate link to a
full Accessibility checklist. Then we can make it a normal required
part of peer review.