Move from master to main in a few places
This DOES NOT change any content of markdown files (other
PRs will be about that).

This is part of https://tracker.moodle.org/browse/MDLSITE-7418

Instead, it just changes the few places were we use "master", namely:
- In the versions drop down in versioned pages like https://moodledev.io/docs/apis/commonfiles/version.php
- In the versions page: https://moodledev.io/versions
- In the scripts in charge of fetching, storing and using the lib/components.json file.
- Moving data/master uses to data/main.

I only have been able to test the 2 first points in netlify, but I think that
the other changes are safe too (it's a simple replacement and move).