Link to the statement of conformance file
WebKey IT used to list its clients' accreditations on its website.
However, since it was acquired by AbleDocs and later became
GrackleDocs, it no longer seems to update this list.

Because of this, the link to the "WCAG 2.1 Level AA conformance" in
our dev docs points to an old accreditation. I think it's better to
replace the link with a URL to the actual statement of conformance
that is currently on the 4.2 accessibility audit epic (MDL-78185).