How to handle privacy providers that only store data in a sub-system
It appears that this is not clear to everyone who reads the documentation at
the moment that a plugin that sends data to a subsystem, for example the
logs should not use the null_provider and that the sub-system will do the
retrieval and deletion of the personal data.

This has been the case in all versions of Moodle supported by the dev docs.