Date/time value fixes
It looks like the format placeholders are those used by strftime() not date(), for example "##today##%d %B %Y##" not "##today##d F Y##".

Also, fixed a broken link.

Also, removed the spaces around the '##' because the three example values at the end of this section don't use spaces which makes it confusing. It appears it doesn't matter if spaces are used by I think we should be consistent here.